WIP factor out parts of smart_cards_screen
