+++ libs/ui/src/index.ts
@@ -13,6 +13,7 @@ export * from './button';
 export * from './button_list';
 export * from './callout';
 export * from './card';
+export * from './card_illustration';
 export * from './cast_vote_records';
 export * from './checkbox_group';
 export * from './checkbox_button';
@@ -71,6 +72,7 @@ export * from './select';
 export * from './session_time_limit_tracker';
 export * from './setup_card_reader_page';
 export * from './set_clock';
+export * from './smart_cards_screen';
 export * from './tabbed_section';
 export * from './table';
 export * from './test_mode';
